image-morph-py/requirements.txt
array_api_strict==2.2
astropy==7.0.0
av==14.0.1
backports.statistics==0.1.0
charset_normalizer==3.3.1
checks==0.2
Cython==3.0.11
defusedxml==0.7.1
hypothesis==6.123.7
imagecodecs==2024.12.30
imageio_ffmpeg==0.4.9
importlib_metadata==8.5.0
imread==0.7.6
ipython==8.12.3
itk==5.4.0
jax==0.4.38
lxml==4.9.3
matplotlib==3.8.1
mpmath==1.3.0
mtrand==0.1
mypy==1.7.1
ndonnx==0.9.3
numpydoc==1.8.0
olefile==0.47
osgeo==0.0.1
pandas==2.2.3
pickle5==0.0.12
pillow_heif==0.21.0
platformdirs==4.3.6
pooch==1.8.2
psutil==5.9.6
pyamg==5.2.1
pydot==3.0.4
pygraphviz==1.14
pyinstaller==6.7.0
PyQt5==5.15.11
PyQt6==6.8.0
PyQt6_sip==13.6.0
PySide6==6.7.0
PySide6==6.8.1.1
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
pytest==7.4.4
pytest_fail_slow==0.6.0
pytest_run_parallel==0.3.0
pytest_timeout==2.3.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.6.0
scikits.talkbox==0.2.5
scipy_doctest==1.6
setuptools==68.2.2
SimpleITK==2.4.0
Sphinx==8.1.3
sympy==1.12
threadpoolctl==3.5.0
typing_extensions==4.12.2
zarr==2.18.4
zstd==1.5.6.1
